1. Cohorts page: 1.1 Moved add cohort form into a AddCohort.jsx dialog. 1.2 Added buttons on Cohorts page to let the admin decide how to create a cohort: manually or through slack. 1.3 Modified loading state logic on add Cohort
