+++ src/pages/Admin/Cohorts/Cohorts.jsx
@@ -3,14 +3,9 @@
     =  THIRD PARTY LIBRARIES =
     ==========================
 */
-import { Box, Container, Paper, Typography, styled } from "@mui/material";
-import {
-  CalendarMonthRounded,
-  LaptopRounded,
-  SchoolRounded,
-} from "@mui/icons-material";
+import { Box, Container, Paper, Typography } from "@mui/material";
+import { Class, CloudDownloadRounded } from "@mui/icons-material";
 import useAxiosPrivate from "../../../hooks/useAxiosPrivate";
-import dayjs from "dayjs";
 /*
     ==========================
     =     REACT LIBRARIES    =
@@ -36,55 +31,20 @@ import styles from "./Cohorts.module.css";
     ==========================
 */
 import AuthFormControl from "../../../components/FormControl/AuthFormControl";
-import FormTextField from "../../../components/TextField/FormTextField";
-import FormSelect from "../../../components/Select/FormSelect";
 import AppButton from "../../../components/Button/AppButton";
 import AppDataGrid from "../../../components/DataGrid/AppDataGrid";
-import AppDatePicker from "../../../components/DatePicker/AppDatePicker";
 import CohortsActions from "./Actions/CohortsActions";
 import Loader from "../../../components/Loader/Loader";
-/*
-    ==========================
-    =     AUX VARIABLES      =
-    ==========================
-*/
-const classList = [
-  "Intro to programming",
-  "React.js",
-  "Node.js/Express",
-  "Ruby on Rails",
-];
+import AddCohort from "./Actions/AddCohort";
 
 const Cohorts = () => {
   /*
-        ==========================
-        =         STATES         =
-        ==========================
-    */
-  const [reset, setReset] = useState(false);
-  const [className, setClassName] = useState("");
-  const [startDate, setStartDate] = useState(dayjs());
-  const [endDate, setEndDate] = useState(dayjs());
-  const [loading, setLoading] = useState(false);
-  const [formError, setFormError] = useState({
-    cohortError: {
-      error: false,
-      errorMessage: "Please enter a valid name",
-    },
-    classNameError: {
-      error: true, //Initial value is blank, this is why I set the error to true.
-      errorMessage: "Please select a class for this cohort",
-    },
-    startDateError: {
-      error: false,
-      errorMessage: "Please select a start date for this cohort",
-    },
-    endDateError: {
-      error: false,
-      errorMessage: "Please select an end date for this cohort",
-    },
-  });
+    ==========================
+    =         STATES         =
+    ==========================
+  */
   const [cohorts, setCohorts] = useState([]);
+  const [loading, setLoading] = useState(false);
   const columns = [
     { field: "id", headerName: "ID", maxWidth: 130, flex: 1 },
     {
@@ -133,61 +93,23 @@ const Cohorts = () => {
       ),
     },
   ];
+  // Dialog states
+  const [openNewCohortDialog, setOpenNewCohortDialog] = useState(false);
   /*
-        ==========================
-        =         HOOKS          =
-        ==========================
-    */
+    ==========================
+    =         HOOKS          =
+    ==========================
+  */
   const axiosPrivate = useAxiosPrivate();
   const navigate = useNavigate();
   const location = useLocation();
   const { setAuth } = useAuth();
-  /*
-        ==========================
-        =     ASYNC FUNCTIONS    =
-        ==========================
-    */
-  const postCohorts = async (newCohort) => {
-    console.log(newCohort);
-    try {
-      const response = await axiosPrivate.post("/cohort", newCohort);
-      return response;
-    } catch (error) {
-      if (error.response.status === 403) {
-        console.error(error);
-        //User is required to validate auth again
-        navigate("/login", { state: { from: location }, replace: true });
-        setAuth({
-          userId: "",
-          userName: "",
-          userEmail: "",
-          role: [],
-          loggedIn: false,
-          avatarUrl: "",
-          isActive: undefined,
-          accessToken: "",
-        });
-      } else {
-        console.error(error);
-      }
-    }
-  };
-
-  /* 
-        ==========================
-        =        EFFECTS         =
-        ==========================
-    */
-  useEffect(() => {
-    setReset(false);
-  });
 
   useEffect(() => {
     let isMounted = true;
     const controller = new AbortController();
 
     const fetchCohorts = async () => {
-      setLoading(true);
       try {
         const response = await axiosPrivate.get("/cohort", {
           signal: controller.signal,
@@ -204,12 +126,10 @@ const Cohorts = () => {
         });
         console.log(formattedCohorts);
         isMounted && setCohorts(formattedCohorts);
-        setLoading(false);
       } catch (error) {
         console.error(error);
         if (error.response.status === 403) {
           //User is required to validate auth again
-          setLoading(false);
           navigate("/login", { state: { from: location }, replace: true });
           setAuth({
             userId: "",
@@ -222,7 +142,6 @@ const Cohorts = () => {
             accessToken: "",
           });
         } else {
-          setLoading(false);
           console.error(error);
         }
       }
@@ -235,100 +154,6 @@ const Cohorts = () => {
       controller.abort();
     };
   }, []);
-  /*
-        ==========================
-        =         HANDLERS       =
-        ==========================
-    */
-  //1. Cohort errorHandler:
-  const handleCohortNameError = (inputError) => {
-    setFormError((prevState) => ({
-      ...prevState,
-      cohortError: {
-        ...prevState.cohortError,
-        error: inputError,
-      },
-    }));
-  };
-  //2. Class name handler
-  const handleClassNameChange = (selectedClassName) => {
-    setClassName(selectedClassName);
-    setFormError((prevState) => ({
-      ...prevState,
-      classNameError: {
-        ...prevState.classNameError,
-        error: false,
-      },
-    }));
-  };
-  //3. Start date handler
-  const handleStartDateChange = (newStartDate) => {
-    setStartDate(newStartDate);
-    setFormError((prevState) => ({
-      ...prevState,
-      endDateError: {
-        ...prevState.endDateError,
-        error: endDate < newStartDate ? true : false,
-      },
-    }));
-  };
-  //4. End date handler
-  const handleEndDateChange = (newEndDate) => {
-    setEndDate(newEndDate);
-    setFormError((prevState) => ({
-      ...prevState,
-      startDateError: {
-        ...prevState.startDateError,
-        error: newEndDate < startDate ? true : false,
-      },
-      endDateError: {
-        ...prevState.endDateError,
-        error: newEndDate < startDate ? true : false,
-      },
-    }));
-  };
-  //5. Form onSubmit event handler
-  const handleCohortSubmit = async (event) => {
-    setLoading(true);
-    event.preventDefault();
-    const newCohort = {
-      name: event.target.cohort.value.trim(),
-      start: startDate.format(),
-      end: endDate.format(),
-      type: className,
-    };
-    const errors = Object.values(formError);
-    try {
-      if (!errors.some((error) => error.error === true)) {
-        const response = await postCohorts(newCohort);
-        console.log(response);
-        setLoading(false); // Stop loading in case of error
-        if (response.status === 201) {
-          console.log(cohorts);
-          setCohorts((prevCohorts) => [
-            ...prevCohorts,
-            {
-              id: response.data.cohort._id,
-              cohort: response.data.cohort.name,
-              class: response.data.cohort.type,
-              startDate: new Date(response.data.cohort.start),
-              endDate: new Date(response.data.cohort.end),
-            },
-          ]);
-        }
-      } else {
-        console.error("Form validation is not letting form submission");
-      }
-    } catch (error) {
-      console.error(error);
-    } finally {
-      setReset(true);
-      setClassName("");
-      setStartDate(dayjs());
-      setEndDate(dayjs());
-    }
-    setLoading(false);
-  };
 
   return (
     <Container maxWidth="lg">
@@ -361,111 +186,26 @@ const Cohorts = () => {
           {" "}
           COHORTS MANAGEMENT{" "}
         </Typography>
-        <Box
-          component={"form"}
-          sx={{
-            display: "flex",
-            flexDirection: "column",
-            justifyContent: "center",
-            alignItems: "center",
-            width: "100%",
-          }}
-          autoComplete="off"
-          onSubmit={handleCohortSubmit}
-        >
-          <div className={styles.formContainer}>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <SchoolRounded fontSize="large"></SchoolRounded>
-                <br></br>
-                <br></br>
-              </Box>
-              <FormTextField
-                required
-                type="text"
-                label="Cohort:"
-                name="cohort"
-                isFocused={true}
-                width="100%"
-                variant="light"
-                regex={/^[a-zA-Z]+( [a-zA-Z]+)*$/}
-                onHandleError={handleCohortNameError}
-                errorMessage={"Please enter a valid name"}
-                reset={reset}
-              ></FormTextField>
-            </AuthFormControl>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <LaptopRounded fontSize="large" />
-                <br></br>
-              </Box>
-              <AuthFormControl
-                width="100%"
-                isNested={true}
-                error={formError.classNameError.error}
-              >
-                <FormSelect
-                  id={"class"}
-                  name={"class"}
-                  label={"Class:"}
-                  selectValue={className}
-                  onSelectValue={handleClassNameChange}
-                  list={classList}
-                  variant={"light"}
-                  multiple={false}
-                  error={formError.classNameError}
-                ></FormSelect>
-              </AuthFormControl>
-            </AuthFormControl>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <CalendarMonthRounded fontSize="large" />
-                <br></br>
-              </Box>
-              <AppDatePicker
-                id={"startDate"}
-                name={"startDate"}
-                label={"Start date:"}
-                dateValue={startDate}
-                onDateValueChange={handleStartDateChange}
-                variant={"light"}
-              ></AppDatePicker>
-              <AppDatePicker
-                id={"endDate"}
-                name={"endDate"}
-                label={"End date:"}
-                dateValue={endDate}
-                onDateValueChange={handleEndDateChange}
-                minDate={startDate}
-                variant={"light"}
-              ></AppDatePicker>
-            </AuthFormControl>
+        <div className={styles.formContainer}>
+          <AuthFormControl width="50%">
             <AppButton
               text={"Add new cohort"}
-              type="submit"
-              width="25%"
+              type="button"
+              width="100%"
+              handlerFunction={() => setOpenNewCohortDialog(true)}
+            >
+              <Class fontSize="large"></Class>
+            </AppButton>
+            <AppButton
+              text={"Add cohort from Slack"}
+              type="button"
+              width="100%"
               handlerFunction={() => {}}
-            />
-          </div>
-        </Box>
+            >
+              <CloudDownloadRounded fontSize="large"></CloudDownloadRounded>
+            </AppButton>
+          </AuthFormControl>
+        </div>
         {loading ? (
           <Box
             sx={{
@@ -477,15 +217,20 @@ const Cohorts = () => {
             <Loader />
           </Box>
         ) : (
-          <Box sx={{ display: "flex" }}>
-            <AppDataGrid
-              columns={columns}
-              rows={cohorts}
-              fieldToBeSorted={"class"}
-              sortType={"asc"}
-            />
-          </Box>
+          <AppDataGrid
+            columns={columns}
+            rows={cohorts}
+            fieldToBeSorted={"class"}
+            sortType={"asc"}
+          />
         )}
+        {openNewCohortDialog ? (
+          <AddCohort
+            open={openNewCohortDialog}
+            handleOpen={setOpenNewCohortDialog}
+            onRegisterCohort={setCohorts}
+          ></AddCohort>
+        ) : null}
       </Paper>
     </Container>
   );
